zephyr: switch to platform/ec printf format implementation
Prior to this change, the cprints and cprintf shim implementations
used Zephyr's printk to do the output formatting.

Our EC code has some custom printf specifiers not supported by
Zephyr's printk.  We've already attempted to send some of our custom
specifiers upstream, but upstream does not want them:

https://github.com/zephyrproject-rtos/zephyr/pull/28882

The logical thing to do would be to bring in the vfnprintf function
from our EC to the Zephyr build, and use that to do the output
formatting instead.  That's what this CL does.

The binary cost of brining in this second printf implementation
appears to be minimal (952 bytes on volteer).

BUG=b:177065615
BRANCH=none
TEST=on posix-ec and volteer, run gettime and observe output no longer
     contains %.6lld, but instead the correct system time

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I53cd4edf129223c12a2c5e7d0519623a8d07a328