zephyr: Define CONFIG_HAS_TASK_PD_INT if needed
ECOS uses this option to indicate whether any PD_INT (Power Delivery
Interrupt) task is running. We must define it the same way for Zephyr
so that the interrupt handler runs as expected.

Define it if the PD_INT port 0 is active, since that means there is at
least one task active.

BUG=b:175434113
BRANCH=none
TEST=make BOARD=volteer -j30
With a zephyr-chrome CL, build volteer on zephyr

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I8bee88211caa34ca1b95474b06b66921bbea23aa