TCPMv2: Defer initializing type-C supplier current limit
Initializing all of the suppliers seeds the charge manger and a
supplier is then selected (a corresponding charger input current
is configured).

Originally, both PD and type-C suppliers are unconditionally
initialized in the restart_tc_sm() that seeds the charge manger.
A lower current supplier (the default VBUS supplier) is selected
first before the type-C current detection.

A issue happens on the first boot from battery cutoff, the EC
RO detects 5V 3A and boots AP. When sysjump to RW, the default
VBUS supplier (5V 0.5A) is used first and limits the charger input
current to some lower value. It makes AP out of power.

This change defers the type-C supplier initialization to either
Unattached.SNK (new) or Attached.SNK.

In Unattached.SNK, we debounce the CC open status to make sure
CC open and then initialize the type-C supplier. As some TCPC
needs time to get the CC status valid. Before that, CC open is
reported by default. It misleads the decision.

We reuse the existing role toggle timer to debounce. And also
align the transitions to DRPAutoToggle and LowPowerState to
simplify the logic.

BRANCH=Trogdor
BUG=b:174105232, b:175663604
TEST=Verified the first boot from cutoff, AP still ON after sysjump
to RW.
TEST=Verified type-C detection, PD detection, and BC 1.2 detection
still work correctly.

Change-Id: Ia5d714a15d3f556463cf38e11130151e1f1477cb
Signed-off-by: Wai-Hong Tam <waihong@google.com>