pd: fix RX re-enabling with the integrated PHY
When using the TCPC 'stub' to connect the TCPMv1 stack to the integrated
PD PHY, we were already re-enabling the RX after a Hard-reset but the
rx_enabled variable was not set in the TCPC code.
Update the stack code to call the rx_enable method as we already do with
regular TCPC, this ensures that the rx_enabled variable is always properly
set.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=b:168621142
BRANCH=servo
TEST=manual, send a hard-reset from the servo v4.1 and see the rest of
the sequence from the DUT happening properly.

Change-Id: I5b8c11d6778c40bb980f69ab0bb4c1619220e28e