pompom: change panel blacklight PWM frequency to 20k
Noise is related with LCD Panel and Brightness PWM.
We reduced the frequency to 200hz to eliminate the noise problem.
But PWM frequency 200Hz is too low will cause the  flicker problem.
Can easily see it when move my eyes around the screen.
Therefore, changing the panel black light PWM frequency to 20kHz can eliminate the flicker problem and the noise still not be heard

BUG=b:175795667
BRANCH=none
TEST=
Boot into OS can still not hear the noise.
Change the PWM frequency from 200hz to 20kHz,there will be no flickering problem on the screen.

Signed-off-by: johnwc_yeh <johnwc_yeh@compal.corp-partner.google.com>
Change-Id: I0e60f6f200f70a69e2970dfb0b5befb494c2b64e