TCPMv2: Do not check for SinkTxOk in FRS-Mode
After an FRS signal, the initial sink will attempt to initiate the FR
Swap AMS, and the initial source should apply SinkTxOk. If the initial
source is removed entirely, FRS will be triggered butSinkTxOk will
never be applied and this will get the initial sink state machine stuck.

So, in the case of an FRS ignore the CC voltage. After an FRS signal,
the source isn't allowed to initiate an AMS.

BUG=b:171740860
TEST=On Volteer with the PS8815, enable FRS and remove the FRS device.
Make sure we enter ErrorRecovery instead of hanging in
PE_PRS_SNK_SRC_Send_Swap
TEST=make buildall
BRANCH=none

Change-Id: I8373a76c0c19feeb909b0623a1ae2d6b0ad5fa60
Signed-off-by: Eric Herrmann <eherrmann@chromium.org>