raa489000: Add support for linear charge
This commit adds support to enable linear charging of the battery.
This leverages the Dynamic Voltage Compensation (DVC) mode of the
RAA489000.  It allows the BFET to be finely controlled to regulate the
precharge current entering the battery pack when charging via the sub
board.

BUG=b:173530987
BRANCH=dedede
TEST=Build and flash magolor.  Deplete battery to 0% such that the
DFET is disabled.  Plug in a charger on the sub board, verify that DUT
is able to revive battery pack and charge the battery at its desired
precharge current rate.
TEST=Cutoff battery, plug in charger on sub board, verify that
battery is able to be revived and charged.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: I364f1014ec00e877fd7f2dfd30ebb0032fd88ffc