max14637: Switch should not be kept open when PD adapter is disconnected
Root Cause:
D+ and D- of PD adapter are shorted so chip would detect PD adapter as
the DCP. As a result, chip would set USB 2 switch open based on product
spec. Later on once USB 2.0 storage is attached, DUT can't recognize it
because USB 2.0 data path is blocked by chip now before introducing this
CL.

Solution:
Whenever adapter or USB client device is disconnected from a port of DUT,
we re-trigger the bc12_detect() so chip will be off then on for detecting
the status again. In this case, the D+/D- are NC so chip will detect it
as the SDP and keep this status afterward.

When USB 2.0 storage is connected later, bc12_detect will not be called
again due to DUT is in source role. At this moment, USB switch is closed
so USB 2.0 path is good. And there is no BC12 detecting cycle happened
so we will not hit issue resolved in CL:*2364342 as well.

When adapter is connected again, bc12_detect will be triggered for
detecting DCP / SDP / CDP.

BUG=b:177265749
BRANCH=octopus
TEST=make buildall -j 8
TEST=check lsblk , usb2 device is exist after reproduce step
Signed-off-by: yu-an.chen@quanta.corp-partner.google.com

Change-Id: I769e9f97daf86992259d8da0bbb38a1068bd8a5a