[WIP] vilboz: cold EC reboot can cause tic-tic-tic
performing a cold reboot on the EC when using a
servo_v4 can cause the EC to almost brownout on
reboot but does not quite get there. This causes
the hardware to perform the tic-tic-tic sounds
with servo_v4 continuing to provide Vbus.

I am adding ideas in this CL in order for a
group collaboration to happen

BUG=b:174526198
BRANCH=zork
TEST=batteryless reboot_ec cold when powered by servo_v4

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: I2dec98d6c9add357685da725fe2d65050ac40919