Refactor CONFIG_FLASH_SIZE to match Zephyr
In Zephyr CONFIG_FLASH_SIZE is a Kconfig value that is used
throughout. The issue is that the units don't match. In
Zephyr the value is in KiB instead of bytes. This refactor
simply changes the units of CONFIG_FLASH_SIZE in platform/ec.
To match Zephyr's so that shimming flash.c will work later on.

Long term this will help as flash size will only be defined per
chip and image offsets/sizes will be computed during the build.

BRANCH=none
BUG=b:174873770
TEST=make buildall
be generated by the build instead of per board

Signed-off-by: Yuval Peress <peress@chromium.org>
Change-Id: I44bf3c7a20fcf62aaa9ae15715be78db4210f384