volteer: Move USB-A charging info to usbc_config
Move this information into a file that can be built by Zephyr, so that
it can enable USB-A charging.

Note: This is not related to USB-C. Should we create a charger.c file
in board/volteer?

BUG=b:175434113
BRANCH=none
TEST=make BOARD=volteer -j30
With a zephyr-chrome CL, build volteer on zephyr
Type 'usbchargemode 0 on' on volteer and see it start charging the phone

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I6658ebba1e473cafed2c8ae05e4013207f26ef3d