usb_mux: Correction in sending the Host event IRQ=0
Retimer is the first mux and the virtual mux is the second Mux.
Here the check should be for virtual Mux (second mux) instead
checked for the first Mux. Because of that there is no IRQ=0
host events sent.
Removed the condition check as it is taken care in side the
usb_mux_hpd_update function.

BUG=b:177876072
BRANCH=None
TEST=HDMI MST HUB second port display also working

Signed-off-by: Madhusudanarao Amara <madhusudanarao.amara@intel.corp-partner.google.com>
Change-Id: I36205f2d442e9dcaf3d082151201e4c210878d8d
Signed-off-by: Madhusudanarao Amara <madhusudanarao.amara@intel.corp-partner.google.com>