dts: binding: add an explicit binding for named-gpios
Currently the name-gpios node uses the gpio-keys binding. That is a
standard binding used in Zephyr for GPIO keys, as in physical button.

This copies the binding over to a dedicated "named-gpios" one, and adds
a #gpio-cells property so that we can refer to its nodes from a phandle
property of a different node.

BUG=none
TEST=build and check generated files
BRANCH=none

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: Iadd3465b23b89cfe08c96c4d03479f1e5f8e3c59