TCPMv2: TCPMv2 Send message with timed response should hard reset within range
PE_SRC_Send_Capabilities is supposed to wait for 24ms before
sending HardReset but that reset must be sent before 30ms.
Setting the delay to 30ms will guarantee we never are within
this range and will always be after it.  TD.PD.SRC.E5 is
guaranteed to fail with this set to 30ms

BUG=b:177999668
BRANCH=zork
TEST=make run-usb_tcpmv2_compliance

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: I8d743680b62e5dea840c935f06eb987c8358e473