npcx7: Set maximum image size to 1/2 flash size
The EC firmware image size is limited by the minimum of 1/2 the flash
size or the code RAM size.

When 1/2 flash is smaller than code RAM, the NPCX was reserving one
flash block for the NPCX header. This has the side effect of reducing
the erase block size from 64 KiB to 4 KiB, and introduces a boot time
regression when processing flash blocks.

CL:2334389 changed to the linker script to account for the NPCX header,
so the NPCX code no longer needs to reserve a flash block to prevent
overflowing flash.

Verified BIOS boot time is improved with this change:
Without change:
 517:waiting for EC to allow higher power draw         580,251 (40,517)

With change:
 517:waiting for EC to allow higher power draw         534,768 (4,859)

BUG=b:175115527
BRANCH=volteer
TEST=make buildall
TEST=Boot EC on Delbin with NPCX797FC
TEST=Verify build fails if RO image size exceeds 256 KiB - 64 bytes.

Signed-off-by: Keith Short <keithshort@chromium.org>
Change-Id: Ide6664165e84da7b88703f2f9a979c7af672a630