[DISCARD] servo_v4p1: PD PHY TX configuration for 'DTS' polarities
Add configuration of the PD TX PHY in the STM32 hardware
to transmit on both CC when in DTS mode (ie POLARITY_CCx_DTS).

Note: the POLARITY_CCx_DTS special polarities should never reach the PHY
code (ie they are filtered before by polarity_rm_dts() in the TCPC), so
this is likely useless in the current situation.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=none
BRANCH=servo
TEST=none

Change-Id: I0699db2d54ccf0337b9fd67beaee6c5e36b8df35