CoachZ: Pull down WLC IRQ
ctn730 push-pulls the IRQ line. When it's off, the line should stay
low to prevent the EC from getting false IRQ.

This patch also configures the ctn730 reset line as GPIO_ODR_HIGH so
that the line is kept high and a reset can be issued if needed.

BUG=b:173235954, b:178096436
BRANCH=trogdor
TEST=CoachZ

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I0023d9a980826a6ea4741459a0b65c2cc11e6106