npcx7: Set the erase size to 64 KiB
The NPCX flash layout was setting the erase size based on RO image size
to ensure the host can erase the entire image using.

When 1/2 flash is smaller than code RAM, the NPCX reserves one 4 KiB
flash block for the NPCX header. This has the side effect of reducing
the erase block size from 64 KiB to 4 KiB, and introduces a boot time
regression when processing flash blocks.

CL:2325764 modified the common flash support to allow the host to
program up to CONFIG_WP_STORAGE_SIZE bytes for the RO image and
CONFIG_EC_WRITABLE_STORAGE_SIZE bytes for the RW image.

Set the erase size to a fixed value of 64 KiB, with build checks
to verify the total flash size is a multiple of 64 KiB.

Both the RO and RW image starting addresses are already aligned to at
least a 64 KiB boundary, so there is no impact when the RW firmware
is upgraded.

BUG=b:175115527
BRANCH=volteer
TEST=make buildall
TEST=Boot EC on Delbin with NPCX797FC
TEST=Increase the RW image size to 0x3ed1c (maximum is 0x3F000) and
verify EC software sync is successful at updating the image.

Signed-off-by: Keith Short <keithshort@chromium.org>
Change-Id: I8823da2b909dfa36646b411a146001c618ce0e8c