zephyr: Add support for BOARD_VERSION
Add Kconfig options for this feature.

At present system_get_board_version() is not built for Zephyr. We could
split it out but it seems better to wait until we have the existing file
brought into the build in toto.

BUG=b:178124020
BRANCH=none
TEST=make BOARD=volteer -j30
boot zephyr for volteer and see that the options are defined

Change-Id: I100c8f3f65db437fbe654e0230e07f59c6e0c7b4
Signed-off-by: Simon Glass <sjg@chromium.org>