zephyr: Bring in common/peripheral.c
This file is intended to be in the core build. Bring it into the Zephyr
build. Also define CONFIG_CROS_BOARD_INFO, used in that file, since it
should depend on CONFIG_PLATFORM_EC_CBI.

BUG=b:178124020
BRANCH=none
TEST=make BOARD=volteer -j30
boot zephyr for volteer and see that it brings in the new code

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: Ie8dd7a74b05c23a5d04781fb2d9bfb1a1cfe1866