elan_private: Implement elan fp sensor on stm32f4
This patch implements the ELAN FP API, which is used to control the
ELAN FP sensor and matching algorithm on stm32f4.

Therefore, we reduce the size of TEMPLATE_SIZE on STM32f4 and
implements elan sensor reset API.

BRANCH=None
BUG=None
TEST=We build on bloonchipper and dartmonkey, and testing Elan sensor
with libelan_515_m4/m7.a and libelan_80_m4/m7.a successfully.

Signed-off-by: herman lin <herman.lin@emc.com.tw>
Change-Id: Iaf4b85744a49a3ae12f20d91740515b7dc198e56