zephyr: test_util: improve EC compatibility
When building a unit test that compiles for either EC or Ztest, the
TEST_MAIN macro should call test_reset and test_print_results as
part of running the tests. This allows the test code to omit the
functions (which are not defined when building for Ztest).

Also add the format parameter so that when testing for floating-point
values being near (zassert_within or TEST_NEAR) we can specify the
format for the EC assertion to print the result.

BUG=b:172240633
BRANCH=None
TEST=`make runhosttests` and `zmake testall`

Signed-off-by: Paul Fagerburg <pfagerburg@google.com>
Change-Id: I23dc933ed4ba0f386fcae68d257161ffffe226eb