zephyr: volteer: make compiling with features disabled easier
Many of the board configuration files depend on certain Kconfig
features, but this dependency was not reflected in the CMakeLists.txt
or the interrupts we enable.

This allows compiling with USB-C, CBI, charging, battery, or power
sequencing disabled, which is useful as a debugging technique to see
if one feature is interfering with another.

BUG=none
BRANCH=none
TEST=compile with CBI, USB-C, and battery disabled, clean compile and
     things work

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I687805d9649badb15d1f2d099f41521f8da10197