zephyr: volteer: Fix build
There seems to be a few issues with the latest changes for the
jhl8040r. Looking at the generated devicetree_unfixed.h the
following properties were missing:
- reset-gpios
- int-gpios
- ls-en-gpios

This can be seen by looking for:
  DT_N_S_soc_S_io_i2c_ctrl3_port0_S_jhl8040r_40_P_.*

For those unfamiliar with how devicetree names these, we're
looking for a concatination of the path leading to the device.
In this case the jhl8040r is under soc/io_i2c_ctrl3_port0.
Which has the label i2c3_0 (this can be found in npcx7m6fb.dtsi
under the Zephyr repo). Then P_ is the prefix used for the
properties.

Since the intel,jhl8040r.yaml was including i2c-device.yaml
(again under the Zephyr repo). This specified on-bus: i2c which
the parent i2c3_0 did not define (nor do I think it should).
Since i2c-device.yaml doesn't declare any new properties, it
suffices to replace it with the inclusion of base.yaml and
setting reg and label to required there.

Finally, another build error was around the addressing of the
jhl8040r. The register only specified a single address while
devicetree was expecting 2 address cells and a single size,
see error below:
    devicetree error: 'reg' property in <Node
    /soc/io_i2c_ctrl3_port0/jhl8040r@40 in 'volteer.dts.pre.tmp'>
    has length 4, which is not evenly divisible by 12
    (= 4*(<#address-cells> (= 2) + <#size-cells> (= 1)))

This is fixed by manually specifying the #address-cells to 1,
and #size-cells to 0.

BRANCH=none
BUG=none
TEST=zmake testall
TEST=manually verify the content of devicetree_unfixed.h

Signed-off-by: Yuval Peress <peress@chromium.org>
Change-Id: Ie2080d0f652fe9a05b3b595927ebba4e30b42e32