vilboz: enable USB 3 mux for the LTE modem
On Vilboz360, if an internal LTE modem is present, it is connected on
the USB C1 port. In this case we need to ensure that the superspeed mux
of the C1 port is in USB 3 position rather than in 'safe mode' where the
LTE modem would enumerate only as USB 2.0 HS since the superspeed pairs
would be disconnected.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=b:177389383
BRANCH=zork
TEST=On vilboz360, see the L850-gl LTE modem enumerating as a superspeed
device.

Change-Id: I0f3e100077552a8712b2177b3a066e124e149dab