zephyr: Use Chrome EC's libgcc implementation over the compiler's
gcc's libgcc is built with a compile time notion of what an
architecture's ABI will look like. If that happens to be wrong, you're
out of luck.

Instead, use our own implementation which, while written in assembly
(and as such not as flexible as it could be) is processed with the
right set of flags.

BUG=b:178363068
BRANCH=none
TEST=zephyr boots on kohaku when built with coreboot-sdk.

Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Change-Id: I6d27bb48478081b6c2ff8927734492282e55e898