zork: add a LTE present bit in fw_config
The up-to-date program/zork/program.star has a new LTE_PRESENT bit in
the fw_config.
Define it, so we can use it to steer the USB 3 mux for the LTE modem.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=b:178457388
BUG=b:177389383
BRANCH=zork
TEST=On a vilboz360 LTE, program the CBI as for the next production
batch (e.g. 'sudo ectool cbi set 6 0x20099200 4') then run with the
next CL and see the LTE modem enumerated as USB 3.

Change-Id: I7834aa5bc8a7ba45debbc1f5c20d82f04de8d84b