max14637: call bc12_detect in VBUS off can't update charge as available
Originally bc12_detect function not only detects BC12 status but also
update result to charge manager. Now bc12_detect can be called not only
when VBUS is on but also off because of CL:*2626791. The result is that
charge manager would report charging available even if VBUS is off.

As a result, we split charger manager update from bc12_detect as the single
function so bc12_detect can be called when VBUS is on or off but charger
manager update for availabel charging will be triggered when VBUS is on.

BUG=b:177845650, b:177265749, b:178509655
BRANCH=octopus
TEST=make buildall -j 8
TEST=check online parameter in /sys/class/power_supply can report
     correct value when PD adapter or BC12 charger is plugged in/out.
TEST=check `ectool usbpdpower` can report correct values for PD adapter
     and BC12 charger.

Change-Id: Ifc4f23edb9272177a6b3637b812b86cf9bef2378
Signed-off-by: Marco Chen <marcochen@chromium.org>
Signed-off-by: Marco Chen <marcochen@google.com>