pd_alt_mode_dfp: Default to 4 lanes of DP
Currently our default PD policy states that if dock that supports DP
alt mode also advertises Multi-function preferred, we configure the
superspeed muxes to use 2 lanes for DisplayPort and 2 lanes for
SuperSpeed USB. However, this is done without regard as to what the
actual usage of the SuperSpeed ports may be.

We've historically made this trade off, but it results in degraded
display performance.  Users may be better served if we prioritize
display performance over SuperSpeed USB.

The ideal solution involves adding additional plumbing such that we
can configure the chosen pin configuration from the AP based upon
other heuristics.

BUG=b:178635286
BRANCH=as many as feasible?
TEST=Build and flash DUT, plug in a dock with SuperSpeed USB ports,
verify that 4 lanes of DP are configured.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Ie2d47588012aceb7f13312b6947b885f8f7034a8