dartmonkey: Add flash_fp_mcu support for strongbad
This is a hack to get the script working.

I lowered the stm32mon connect retries to allow the flash_fp_mcu script
the ability to reset the chip on reset.

There also seems to be some gpio issue with boot0, where it needs to be
reasserted before coming out of reset (into bootloader mode).

BRANCH=none
BUG=none
TEST=Cros fingers

Change-Id: I33d8fee21a2afbf8f5c399e6cb47c2095d833f57
Signed-off-by: Craig Hesling <hesling@chromium.org>