dartmonkey: Refactor GPIOs to keep RO minimal
We want the amount of code and number of things defined in RO to be as
minimal as possible since RO is frozen forever. By keeping RO minimal,
we can reduce surface area for attacks and also confusion when GPIOs are
removed or renamed.

The fingerprint-related code only runs in RW, so move all
fingerprint-related GPIOs and associated code into separate files that
are only included in RW.

BRANCH=none
BUG=b:175115925, b:178746753, b:b:177908650
TEST=Flash on icetower v0.1, verify sensor shows up (with SPI_SEL change)
TEST=Flash on nocturne, enroll, lock, unlock

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: Id59d4cd8011012ba4fd6823e1464c661784d4689