usb_pd: Improve safe state when entering DP mode
Place the USB Type-C pins that are to be re-configured to DisplayPort
Configuration into the Safe state before sending the configure command.
Then switch the pins to DisplayPort after receiving the ack to the
command.

This is handled correctly in Zork. Move this handling into the common
code so it applies to all platforms.

BUG=b:123310411, b:178733280
TEST=Functional test of DP over Type-C
BRANCH=None

Change-Id: I29762a0fa8e780cbdf391cde298422d6c01f6f0a
Signed-off-by: Rob Barnes <robbarnes@google.com>