device_event: Don't notify host of disabled events (2)
CL:2659684 made device_set_events skip notification for disabled
flags and flags waiting to be consumed by the host.

Resending notification for pending flags is redundant in most cases
but we still need to do so in case the host didn't have a chance to
consume the flag. Otherwise, the host would be never be notified
thus the flag would never be consumed.

BUG=b:36024430, b:178537669, b:177664326
BRANCH=Trogdor
TEST=CoachZ. Verified host event isn't sent until powerd enables WLC.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I186c2ef0c4ffd3fc9c7fad1cba9e1478669735a2