servo_updater: firmware channel retrieval as a library service
This change introduces the ability of using servo_updater as a library
to check whether a firmware string (e.g. the one currently read out from
the servo device) belongs to a known channel. If so, the channel will be
returned. If not, None is returned to indicate this is an unknown
firmware string. This can then be used to warn the user to update, or
inform them what channel they are currently running.

BRANCH=None
BUG=b:179310743
// This is performed with the CL that uses this: chromium:2661015
TEST=sudo servod -b soraka
     dut-control servo_micro_firmware_channel
servo_micro_firmware_channel:stable

Change-Id: I49e920c7c7977b7c05828f8464af973f538f7aeb
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>