bloonchipper: Refactor GPIOs to keep RO minimal
We want the amount of code and number of things defined in RO to be as
minimal as possible since RO is frozen forever. By keeping RO minimal,
we can reduce surface area for attacks and also confusion when GPIOs are
removed or renamed.

The fingerprint-related code only runs in RW, so move all
fingerprint-related GPIOs and associated code into separate files that
are only included in RW.

BRANCH=none
BUG=b:178746753
TEST=On dragonclaw v0.2 with servo micro and jtrace, verify sensor ID
     shows on console
TEST=On dragonair, test fingerprint enroll and unlock
TEST=On morphius, test fingerprint enroll and unlock

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: Ia6791666b57ed9be81230d8a5f77d874f56da5a4