power: Allow use of custom power sequencer
This allows boards to specify a custom power sequencer. When
CONFIG_CHIPSET_POWER_CUSTOM is defined, the standard chipset power
sequencer code is excluded form the build thus allowing for board
specific code to implement the sequencer functionality.

In addition, the CONFIG_POWER_INTEL_X86 config option is added under the
control of config.h so that power/intel_x86.c is included in the build
as before for intel chipsets when a custom power sequencer is used.

BRANCH=none
BUG=b:173575131,b:177275055,b:177277633
TEST=buildall passes

Change-Id: I04bad2dee0c18f14a2c087467ade600f2c80576d
Signed-off-by: Caveh Jalali <caveh@chromium.org>