TCPMv2: Consult board before setting power role swap request
If a board does not want to power role swap (ex. because it is in S5),
then do not allow the DPM request for the power role swap to be set.
This is equivalent to the logic done when a PR_SWAP message is received
by the system.

BRANCH=None
BUG=b:174726240
TEST=on madoo, plug in phone and shutdown.  Verify DUT does not swap to
sourcing when in S5

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ia6684006d0de6d9f23bac8bd792c34191fabab14
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2676931
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
(cherry picked from commit b3f5b2b0dd38b5bf92eb14aa0a609142b9588f89)