rt1715: Choose Low Power Mode instead of Standby Mode
By default RT1715 TCPC is going to Standby Mode. The TCPC provides
vendor register in which one can enable Low Power Mode. According
to documentation typical current for Standby Mode is 2150uA, for Low
Power Mode it is only 25uA.

When Low Power Mode is enabled and TCPC is looking for connection,
attaching device to USB-C will cause bit 0 in RT_INT vendor register
to be set to 1, which means that TCPC exited Low Power Mode. In
Standby Mode this bit is always set to 0.

BUG=b:176986511
BRANCH=none
TEST=Run EC on Voxel board. Check if USB-C port works correctly.

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: I333b9c782c70c38aed36875991bca4e1a32eaa9a