zephyr: zmake: better parallelization/logging of pytest
Run pytest file-by-file so that we can get better parallelization of
the tests.  While we are add it, add a small handler which parses the
pytest output and prints the relevant info at the right log levels.

This is needed so that we can see what is actually going on with CQ flake.

BUG=chromium:1175647
BRANCH=none
TEST=zmake -l DEBUG testall, observe log levels

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: Ia978ee2985151c0442e2ba74b429c751b0ba9152