zephyr: zmake: delay locating CrOS checkout
For building outside of a ChromeOS environment, a user may have their
own choice of where they put all of the repos for building Zephyr, and
using the ChromeOS directory structure to find the repos in this case
is unwanted.

This also applies to the case where we have an ebuild to build the
sources, cros-workon will locate the sources in different paths.

We delay the evaluation of util.locate_cros_checkout until the code
actually uses it, allowing us to not pass a checkout or locate one
unless we actually need it.

BUG=b:179813084
BRANCH=none
TEST=zmake testall (note nothing useful for getting -m working yet,
     requires follow-on CLs)

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I5e0c6be294d907cf9ed528d2f938303c9fa8f275