tcpmv2: Update the mux state when PD connection established
The entry in Attached.SNK configures the mux and enables the SuperSpeed
USB pairs are required by the PD spec.

This change reconfigures the mux data role once source caps are received
from the partner. This ensures that SuperSpeed USB pairs are disabled
when a PD capable charger, without USB communication, is connected.

BUG=b:178727224
BRANCH=none
TEST=make buildall
TEST=connect non-PD charger, verify mux state enables USB
TEST=connect PD charger, verify mux state disables USB
TEST=connect Gatkex, verify USB4 entry
TEST=conned DP+USB display. Verify USB and DP entry.

Signed-off-by: Keith Short <keithshort@chromium.org>
Change-Id: I16cf82dff7434bddfa867093a6385c19122444db