usb_mux: ps8822: Add support to adjust rx DP EQ setting
This CL modifies the PS8822 usb_mux driver to support multiple pages
(there is a page 0, 1, and 2 for this chip). The usb_mux driver only
needs access to page 0, but page 1 access is needed for DP EQ
settings.

In addition this CL adds a new driver function to allow setting of the
DP Rx equalization gain value.

BUG=b:178656398
BRANCH=None
TEST=tested on quiche and verified the register is programmed when DP
mode is active.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I65cb2fc2c982b01537eb60c0b9da30f4363419aa