zephyr: fix version console command
Fixes the data seen by the version console command. Since the version
data structure is suppose to be at the same offset from the start of
the image, running `version` in the ec terminal previously did not
yield any values for the RW image (the .rodata.ver section was at a
different offset for RW vs RO).

- Add a new fixed-sections.ld file that adds the .fixed section.
- Include it into all our builds at the start of the sections.
- Add a new macro in common.h to allow us to specify that a section
  should be at a fixed location (FIXED_SECTION).
- Use that macro to put the version information in that location.

Sample output:
21-02-09 12:59:45.531 version
21-02-09 12:59:46.638 Chip:    chromeos emu
21-02-09 12:59:46.642 Board:   1
21-02-09 12:59:46.647 RO:      _v2.0.7390+2d340147c
21-02-09 12:59:46.672 RW:      _v2.0.7390+2d340147c
21-02-09 12:59:46.675 Build:   _v2.0.7390+2d340147c
21-02-09 12:59:46.680          2021-02-09 12:58:26 peress@technodrome.c.googlers.com

BRANCH=none
BUG=b:164421798
TEST=build volteer, flash, run `version` command.

Signed-off-by: Yuval Peress <peress@chromium.org>
Change-Id: Id19a07d16b5e307d81e8a411b933d9e9a674eb27