Charge_state_v2: Report discharge when no active charger is present
For the purpose of LED display, report discharge any time there is no
active charge port present.  This may occur in situations where a source
is connected in S5 which we will not charge from (ex. a phone).  The
charge_state_v2 code will consider the state to be charging in these
scenarios, since a Vbus supply is present.

BRANCH=None
BUG=b:174726240
TEST=on galith, shut down with phone attached and note that the charging
LED does not turn on.  Attach a charger and verify LED does turn on.

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I80ce46ae009250eddf3b83398a1e12822d3c5f3c