common: uart: split off host commands to separate module
Split off UART host commands to a separate module for better
modularity and to enable sharing the host command code with Zephyr.

Note: originally I had intended to share the entire uart_buffering
module with Zephyr, but our initial goal was to get working with
printk as the backend instead of Zephyr's UART drivers (they are
rather complex).  We may end up sharing the rest of uart_buffering in
the future, but this lets us scale up/down the amount of UART code
sharing we do as needed.

BUG=b:178033156
BRANCH=none
TEST=make buildall

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: Ib63e46700fb79b46c7ec2ea52dfcbe99d9fd8bbb